+++ .dialyzer_ignore.exs
@@ -1,3 +1,1 @@
-[
-  ~r|lib/mix/tasks/recode.ex:.*:no_return|
-]
+[]